breakout component control switch
